requirements.txt
Authlib==1.2.0
Flask==1.1.2
pandas==1.4.4
pandas_gbq==0.19.1
protobuf==4.22.1
python-dotenv==1.0.0
